# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Система контроля посещаемости и успеваемости",
    "description": "REST API ВСПК: занятия, посещаемость по свайпам/QR, оценки и отчёты.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/teacher/lessons": {
      "get": {
        "tags": [
          "Преподаватель"
        ],
        "summary": "Список занятий преподавателя",
        "description": "Получить список занятий на сегодня или текущую неделю. Параметр period: 'today' или 'week' (по умолчанию 'week').",
        "operationId": "get_lessons_api_v1_teacher_lessons_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "period",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^(today|week)$",
              "default": "week",
              "title": "Period"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LessonWithGroupName"
                  },
                  "title": "Response Get Lessons Api V1 Teacher Lessons Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Преподаватель"
        ],
        "summary": "Создать занятие",
        "description": "Создать новое занятие (пару). teacher_id берётся из токена автоматически.",
        "operationId": "create_lesson_api_v1_teacher_lessons_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LessonCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LessonResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/teacher/lessons/{lesson_id}/students": {
      "get": {
        "tags": [
          "Преподаватель"
        ],
        "summary": "Список студентов группы для пары",
        "description": "Получить список студентов группы, привязанной к указанному занятию.",
        "operationId": "get_lesson_students_api_v1_teacher_lessons__lesson_id__students_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "lesson_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Lesson Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/StudentWithUserInfo"
                  },
                  "title": "Response Get Lesson Students Api V1 Teacher Lessons  Lesson Id  Students Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/teacher/lessons/{lesson_id}/status": {
      "patch": {
        "tags": [
          "Преподаватель"
        ],
        "summary": "Обновить статус занятия",
        "description": "Изменить статус занятия: PLANNED → ACTIVE → FINISHED.",
        "operationId": "update_lesson_status_api_v1_teacher_lessons__lesson_id__status_patch",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "lesson_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Lesson Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LessonStatusUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LessonResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "LessonCreate": {
        "properties": {
          "subject_name": {
            "type": "string",
            "title": "Subject Name",
            "description": "Название предмета"
          },
          "group_id": {
            "type": "integer",
            "title": "Group Id",
            "description": "ID группы"
          },
          "date_time": {
            "type": "string",
            "format": "date-time",
            "title": "Date Time",
            "description": "Дата и время занятия"
          },
          "classroom": {
            "type": "string",
            "title": "Classroom",
            "description": "Аудитория"
          }
        },
        "type": "object",
        "required": [
          "subject_name",
          "group_id",
          "date_time",
          "classroom"
        ],
        "title": "LessonCreate"
      },
      "LessonResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "subject_name": {
            "type": "string",
            "title": "Subject Name"
          },
          "teacher_id": {
            "type": "string",
            "format": "uuid",
            "title": "Teacher Id"
          },
          "group_id": {
            "type": "integer",
            "title": "Group Id"
          },
          "date_time": {
            "type": "string",
            "format": "date-time",
            "title": "Date Time"
          },
          "classroom": {
            "type": "string",
            "title": "Classroom"
          },
          "status": {
            "$ref": "#/components/schemas/LessonStatus"
          }
        },
        "type": "object",
        "required": [
          "id",
          "subject_name",
          "teacher_id",
          "group_id",
          "date_time",
          "classroom",
          "status"
        ],
        "title": "LessonResponse"
      },
      "LessonStatus": {
        "type": "string",
        "enum": [
          "PLANNED",
          "ACTIVE",
          "FINISHED"
        ],
        "title": "LessonStatus"
      },
      "LessonStatusUpdate": {
        "properties": {
          "status": {
            "$ref": "#/components/schemas/LessonStatus"
          }
        },
        "type": "object",
        "required": [
          "status"
        ],
        "title": "LessonStatusUpdate"
      },
...
      "LessonWithGroupName": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "subject_name": {
            "type": "string",
            "title": "Subject Name"
          },
          "teacher_id": {
            "type": "string",
            "format": "uuid",
            "title": "Teacher Id"
          },
          "group_id": {
            "type": "integer",
            "title": "Group Id"
          },
          "date_time": {
            "type": "string",
            "format": "date-time",
            "title": "Date Time"
          },
          "classroom": {
            "type": "string",
            "title": "Classroom"
          },
          "status": {
            "$ref": "#/components/schemas/LessonStatus"
          },
          "group_name": {
            "type": "string",
            "title": "Group Name"
          }
        },
        "type": "object",
        "required": [
          "id",
          "subject_name",
          "teacher_id",
          "group_id",
          "date_time",
          "classroom",
          "status",
          "group_name"
        ],
        "title": "LessonWithGroupName"
      },
...
      "StudentWithUserInfo": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "user_id": {
            "type": "string",
            "format": "uuid",
            "title": "User Id"
          },
          "group_id": {
            "type": "integer",
            "title": "Group Id"
          },
          "student_card_number": {
            "type": "string",
            "title": "Student Card Number"
          },
          "full_name": {
            "type": "string",
            "title": "Full Name"
          },
          "email": {
            "type": "string",
            "title": "Email"
          },
          "avatar_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avatar Url"
          }
        },
        "type": "object",
        "required": [
          "id",
          "user_id",
          "group_id",
          "student_card_number",
          "full_name",
          "email"
        ],
        "title": "StudentWithUserInfo"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/auth/login"
          }
        }
      }
    }
  }
}